backend/requirements/base.txt
Django>=5.1,<7.0
django-environ>=0.11,<1.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.3,<25.0
Pillow>=10.4,<11.0
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
argon2-cffi>=23.1,<24.0
